requirements.txt
numpy==1.24.4
pandas==2.0.3
Pillow==11.1.0
PyYAML==6.0.2
PyYAML==6.0.2
torch==2.4.1
tqdm==4.67.1
